pyproject.toml
[project]
name = "blog-ecological-restoration"
version = "0.1.0"
description = "Add your description here"
readme = "readme.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8",
    "django-autocomplete-light>=3.12.1",
    "django-flatpickr>=2.0.3",
    # "django-googledrive-storage>=1.6.0", # google drive option
    "django-formset>=2.1.4",
    "django-storages>=1.14.6",
    "google-api-python-client>=2.187.0",
    "google-cloud-storage>=3.3.1", # google bucket option
    "gunicorn>=23.0.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
]

[project.scripts]

[dependency-groups]
dev = [
    "djhtml>=3.0.10",
    "djlint>=1.36.4",
    "mypy>=1.19.0",
    "pre-commit>=4.5.0",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "ruff>=0.14.9",
]

[tool.pytest.ini_options]
# django_find_project = false
pythonpath = [".", "blogged"]
DJANGO_SETTINGS_MODULE = "tests.settings"
# addopts = [
#     "--maxfail 1 --reuse-db"
#     # --reuse-db,
#     # "--create-db",
# ]
testpaths = "tests"
markers = ["asyncio: asyncio mark"] # https://github.com/jonghwanhyeon/python-ffmpeg/issues/21
